fix(deployments): fix image source update validation (Issue #1529)
